+++ src/ulong.c
@@ -19,7 +19,7 @@ SEXP R_flint_ulong_initialize(SEXP object, SEXP s_length, SEXP s_x)
 		checkType(s_x, R_flint_sexptypes + 1, __func__);
 		n = (unsigned long long int) XLENGTH(s_x);
 	}
-	ulong *y = (ulong *) ((n) ? flint_calloc(n, sizeof(ulong)) : 0);
+	ulong *y = (ulong *) ((n) ? flint_calloc((size_t) n, sizeof(ulong)) : 0);
 	R_flint_set(object, y, n, (R_CFinalizer_t) &R_flint_slong_finalize);
 	switch (TYPEOF(s_x)) {
 	case NILSXP:
